requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
aiofiles==24.1.0
cryptography==44.0.0
qrcode[pil]==8.0
pyotp==2.9.0
python-magic==0.4.27
pillow==11.0.0
email-validator==2.2.0
httpx==0.28.0  # For API calls and TestClient
pytest==8.3.4
pytest-asyncio==0.24.0
python-dotenv==1.0.1
pillow-heif==0.17.0
zipfile36==0.1.3
# fastapi-csrf-protect==0.1.0  # Incompatible with newer FastAPI versions
slowapi==0.1.9  # For rate limiting

# Database drivers
psycopg2-binary==2.9.9  # PostgreSQL driver for production

# S3 and cloud storage
boto3==1.35.0  # AWS SDK for S3
botocore==1.35.0  # AWS SDK core
jinja2==3.1.4  # Templating engine for email templates